Code seperation of concerns and clean up
