+++ BadgeLedApp/Message/Message+Helper.swift
