+++ BadgeLedApp/Message/Message+Mode+Speed.swift
