+++ BadgeLedApp/Message/Message.swift
