+++ BadgeLedApp/PixelGrid/PixelGrid+Equatable.swift
