+++ BadgeLedApp/PixelGrid/PixelGrid+Scaling.swift
